WithHTMLYouTubeURL/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
jinja2
aiofiles
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib